+++ docs/src/pages/framed/GlobalTheme/GlobalTheme.svelte
@@ -1,6 +1,6 @@
 <script>
   import {
-    Theme,
+    GlobalTheme,
     RadioButtonGroup,
     RadioButton,
   } from "carbon-components-svelte";
@@ -8,7 +8,7 @@
   let theme = "g90";
 </script>
 
-<Theme bind:theme />
+<GlobalTheme bind:theme />
 
 <RadioButtonGroup legendText="Carbon theme" bind:selected="{theme}">
   {#each ["white", "g10", "g80", "g90", "g100"] as value}
