+++ docs/src/pages/framed/GlobalTheme/GlobalThemeToggleCustom.svelte
@@ -1,8 +1,8 @@
 <script>
-  import { Theme } from "carbon-components-svelte";
+  import { GlobalTheme } from "carbon-components-svelte";
 </script>
 
-<Theme
+<GlobalTheme
   render="toggle"
   toggle="{{
     themes: ['g10', 'g80'],
