+++ src/index.js
@@ -112,7 +112,7 @@ export { Tabs, Tab, TabContent, TabsSkeleton } from "./Tabs";
 export { Tag, TagSkeleton } from "./Tag";
 export { TextArea, TextAreaSkeleton } from "./TextArea";
 export { TextInput, TextInputSkeleton, PasswordInput } from "./TextInput";
-export { Theme } from "./Theme";
+export { GlobalTheme } from "./GlobalTheme";
 export {
   Tile,
   ClickableTile,
